# api/swagger/admserver/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Docs",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/system/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System.User"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order asc or desc.",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field.",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ListUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System.User"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "Param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v1/system/users/{name}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System.User"
                ],
                "summary": "Get user info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System.User"
                ],
                "summary": "Update user info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "avatar",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "male",
                            "female",
                            "secret"
                        ],
                        "type": "string",
                        "description": "Gender, male female secret",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "nickname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status, 1-enabled, 2-disabled",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System.User"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "string"
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                }
            }
        },
...
        "v1.CreateUserRequest": {
            "type": "object",
            "required": [
                "countryCode",
                "nickname",
                "password",
                "username"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 130,
                    "minimum": 0,
                    "example": 0
                },
                "avatar": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string",
                    "example": "us"
                },
                "email": {
                    "type": "string",
                    "example": "peter@gmail.com"
                },
                "gender": {
                    "description": "Gender, male female secret",
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "secret"
                    ],
                    "example": "male"
                },
                "nickname": {
                    "type": "string",
                    "example": "Peter"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                },
                "phone": {
                    "type": "string",
                    "example": "9999999999"
                },
                "pid": {
                    "type": "string",
                    "example": "88888888"
                },
                "status": {
                    "description": "Status, 1-enabled, 2-disabled",
                    "type": "integer",
                    "default": 1,
                    "enum": [
                        1,
                        2
                    ]
                },
                "username": {
                    "type": "string",
                    "example": "peter"
                }
            }
        },
...
        "v1.ListUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.UserInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "v1.UserInfo": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "googleStatus": {
                    "type": "string"
                },
                "inviteCount": {
                    "type": "integer"
                },
                "kycStatus": {
                    "description": "KYC status, 0-not verify, 1-pending, 2-verified, 3-failed",
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "payPassword": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "status": {
                    "description": "Status, 1-enabled, 2-disabled",
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}